`refactor: move email-related classes to system package`
