+++ src/main/java/info/mackiewicz/bankapp/presentation/auth/service/UserRegistrationService.java
@@ -1,13 +1,9 @@
 package info.mackiewicz.bankapp.presentation.auth.service;
 
-import java.math.BigDecimal;
-
-import org.springframework.stereotype.Service;
-
 import info.mackiewicz.bankapp.account.model.Account;
 import info.mackiewicz.bankapp.account.service.AccountService;
-import info.mackiewicz.bankapp.notification.email.EmailService;
 import info.mackiewicz.bankapp.presentation.auth.dto.UserRegistrationDto;
+import info.mackiewicz.bankapp.system.notification.email.EmailService;
 import info.mackiewicz.bankapp.transaction.model.Transaction;
 import info.mackiewicz.bankapp.transaction.service.TransactionService;
 import info.mackiewicz.bankapp.user.UserMapper;
@@ -15,6 +11,9 @@ import info.mackiewicz.bankapp.user.model.User;
 import info.mackiewicz.bankapp.user.service.UserService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
 
 @Slf4j
 @RequiredArgsConstructor
